+++ local_planner/src/nodes/planner_functions.h
@@ -57,8 +57,8 @@ void filterPointCloud(
     double min_cloud_size, double min_dist_backoff, double sphere_radius,
     Box histogram_box, const geometry_msgs::Point &position,
     double min_realsense_dist);
-void calculateFOV(double h_FOV, double v_FOV, std::vector<int> &z_FOV_idx,
-                  int &e_FOV_min, int &e_FOV_max, double yaw, double pitch);
+void calculateFOV(float h_FOV, float v_FOV, std::vector<int> &z_FOV_idx,
+                  int &e_FOV_min, int &e_FOV_max, float yaw, float pitch);
 void propagateHistogram(Histogram &polar_histogram_est,
                         pcl::PointCloud<pcl::PointXYZ> reprojected_points,
                         std::vector<double> reprojected_points_age,
